Optimize UTF-8 Int64/UInt64 Formatter for x86 (#25626)
* First attempt to optimize utf8 integer formatter for x86

* Refactoring and code cleanup

* Inlining some methods and adding Int32 specific tests.

* Cleaning up use of ifdef

* Address PR feedback + remove unnecessary casts/using directives.

* Leftover clean up and removing extra new lines

* Fix Debug.Assert in CountDigits

* Fixing nits.

* Reverting compaction using ternary operator.

+++ src/System.Memory/src/System/Number/Number.FormatAndParse.cs
@@ -196,7 +196,7 @@ namespace System
         //
         private static ulong Mul32x32To64(uint a, uint b)
         {
-            return (ulong)a * (ulong)b;
+            return a * (ulong)b;
         }
 
         //
@@ -518,7 +518,7 @@ namespace System
 
             public static unsafe ulong Mantissa(double d)
             {
-                return (ulong)*((uint*)&d) | ((ulong)(*((uint*)&d + 1) & 0x000fffff) << 32);
+                return (*((uint*)&d)) | ((ulong)(*((uint*)&d + 1) & 0x000fffff) << 32);
             }
 
             public static unsafe bool Sign(double d)
